+++ src/actions/index.js
@@ -4,3 +4,5 @@ export * from './appInstance';
 export * from './users';
 export * from './layout';
 export * from './action';
+export * from './canvas';
+export * from './uppy';