+++ src/reducers/index.js
@@ -4,6 +4,8 @@ import appInstanceResources from './appInstanceResources';
 import users from './users';
 import appInstance from './appInstance';
 import layout from './layout';
+import canvas from './canvas';
+import uppy from './uppy';
 
 export default combineReducers({
   // keys should always be lowercase
@@ -12,4 +14,6 @@ export default combineReducers({
   users,
   appInstance,
   layout,
+  canvas,
+  uppy,
 });