+++ src/types/index.js
@@ -4,3 +4,6 @@ export * from './users';
 export * from './appInstance';
 export * from './layout';
 export * from './action';
+export * from './canvas';
+export * from './file';
+export * from './uppy';