+++ CHANGELOG.md
@@ -2,6 +2,17 @@
 
 All notable changes to this project will be documented in this file. See [standard-version](https://github.com/conventional-changelog/standard-version) for commit guidelines.
 
+### [0.1.2](https://github.com/graasp/graasp-app-post-it-notes/compare/v0.1.1...v0.1.2) (2021-06-17)
+
+
+### Features
+
+* add functionality to edit and delete notes ([3f3d67a](https://github.com/graasp/graasp-app-post-it-notes/commit/3f3d67a0c51217484dd664d29738ab608553184d)), closes [#3](https://github.com/graasp/graasp-app-post-it-notes/issues/3)
+* create initial layout ([3ae97e3](https://github.com/graasp/graasp-app-post-it-notes/commit/3ae97e30cd350b0e49a42056351a2b8748ab80e3)), closes [#1](https://github.com/graasp/graasp-app-post-it-notes/issues/1)
+* extract and display user names ([dcee75e](https://github.com/graasp/graasp-app-post-it-notes/commit/dcee75e836497c3daba47c82d9742ce709a8e5f3)), closes [#6](https://github.com/graasp/graasp-app-post-it-notes/issues/6)
+* make notes draggable ([4d46981](https://github.com/graasp/graasp-app-post-it-notes/commit/4d46981884833742969cdd04b9d6639ada493cc5)), closes [#2](https://github.com/graasp/graasp-app-post-it-notes/issues/2)
+* persist notes to db ([42f2382](https://github.com/graasp/graasp-app-post-it-notes/commit/42f2382e8b1b8ba690e0f704966d59c8a3bb1d9e)), closes [#4](https://github.com/graasp/graasp-app-post-it-notes/issues/4)
+
 ### 0.1.1 (2021-06-17)
 
 ## 0.1.0 (2021-04-19)